# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upload Assistant API",
    "version": "2.0.0-dev",
    "description": "Local-first Chinese PT retorrent automation with durable steps, rule gates, and auditable evidence."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/api/v2/jobs/{job_id}/resume": {
      "post": {
        "operationId": "resumeJob",
		"description": "Queues a paused, blocked, or failed job. Unknown target/image upload outcomes and partial or unknown downloader adds require an explicit attempt-bound reconciliation decision; a generic retry is rejected with reconciliation_required.",
        "parameters": [{"$ref": "#/components/parameters/JobID"}],
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResumeJobRequest"}}}},
        "responses": {"200": {"description": "Job queued for recovery", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobEnvelope"}}}}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v2/notification-channels/{name}/probe": {
      "post": {
        "operationId": "probeNotificationChannel",
        "description": "Explicitly send one durable test message. The delivery intent is persisted before the external write, automatic retry is disabled, and an ambiguous result remains reconcilable through the notification record.",
        "parameters": [{"$ref": "#/components/parameters/ResourceName"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["confirm_delivery"], "properties": {"confirm_delivery": {"type": "boolean", "const": true}}}}}},
        "responses": {"200": {"description": "Test notification delivered with a validated receipt"}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"description": "Delivery outcome is unknown and requires reconciliation"}, "502": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v2/metadata-providers/{name}/probe": {
      "post": {"operationId": "probeMetadataProvider", "description": "Explicitly execute one read-only production-contract lookup against a stable public reference. The response returns hashes and call metadata only, never titles or descriptions. Transport, DNS, TLS, timeout, authentication, rate-limit, invalid-response and missing-output failures use safe stable error codes; an unreachable workers.dev PTGen endpoint recommends a reachable custom domain.", "parameters": [{"$ref": "#/components/parameters/ResourceName"}], "responses": {"200": {"description": "Metadata query contract validated"}, "400": {"$ref": "#/components/responses/Problem"}, "429": {"$ref": "#/components/responses/Problem"}, "502": {"$ref": "#/components/responses/Problem"}, "504": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/metadata-providers/{name}/resolve": {
      "post": {"operationId": "resolveExternalMetadata", "parameters": [{"$ref": "#/components/parameters/ResourceName"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MetadataResolveRequest"}}}}, "responses": {"200": {"description": "Normalized IDs, bounded description, and hash evidence"}, "400": {"$ref": "#/components/responses/Problem"}, "502": {"$ref": "#/components/responses/Problem"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "ua_API_token"}
    },
    "parameters": {
      "JobID": {"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
      "ResourceName": {"name": "name", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"}},
...
    },
    "responses": {
      "Problem": {"description": "Structured API problem", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "JobStatus": {"type": "string", "enum": ["draft", "queued", "running", "paused", "blocked", "failed", "complete", "cancelled"]},
...
      "Blocker": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "site_code": {"type": "string"}, "obligation_id": {"type": "string"}},
        "additionalProperties": true
      },
      "NextAction": {
        "type": "object",
        "required": ["action"],
        "properties": {"action": {"type": "string"}, "description": {"type": "string"}, "parameters": {"type": "object", "additionalProperties": true}},
        "additionalProperties": true
      },
...
      "ResumeJobRequest": {
        "type": "object",
        "additionalProperties": false,
		"properties": {"resume_state": {"type": "object", "additionalProperties": true, "default": {}, "properties": {"reconciliation": {"$ref": "#/components/schemas/JobReconciliation"}}}}
      },
	  "JobReconciliation": {
		"type": "object", "additionalProperties": false,
		"required": ["blocker_code", "attempt_id", "decision", "confirmed", "evidence_sha256", "observed_at"],
		"properties": {
		  "blocker_code": {"type": "string", "enum": ["target_upload_outcome_unknown", "image_upload_outcome_unknown", "downloader_partial_add_requires_reconciliation", "downloader_add_outcome_unknown"]},
		  "attempt_id": {"type": "string", "format": "uuid", "description": "Exact latest blocked/failed attempt returned by the attempts API."},
		  "decision": {"type": "string", "enum": ["verified_not_applied", "verified_uploaded", "verified_remote_state"]},
		  "confirmed": {"type": "boolean", "const": true},
		  "evidence_sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$", "description": "SHA-256 of the operator-reviewed external reconciliation evidence; raw evidence and credentials are not copied into the job event."},
		  "observed_at": {"type": "string", "format": "date-time"},
		  "observed_hash": {"type": "string", "description": "Required for a partial or unknown downloader add and must match the exact submitted hash in the persisted reconciliation template."},
		  "observed_torrent_id": {"type": "string", "pattern": "^[0-9]{1,20}$", "description": "Required for verified_uploaded; the worker performs a fresh read-only target search and must observe this exact torrent id."},
		  "submitted_torrent_sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$", "description": "Required for target verified_uploaded and must equal the immutable submitted target torrent SHA-256 in the persisted reconciliation template."},
		  "pending_evidence_sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$", "description": "Required for image verified_uploaded and must equal the server-generated pending image evidence SHA-256 in the persisted reconciliation template."}
		},
		"oneOf": [
		  {"properties": {"blocker_code": {"const": "target_upload_outcome_unknown"}, "decision": {"const": "verified_not_applied"}}},
		  {"required": ["observed_torrent_id", "submitted_torrent_sha256"], "properties": {"blocker_code": {"const": "target_upload_outcome_unknown"}, "decision": {"const": "verified_uploaded"}}},
		  {"properties": {"blocker_code": {"const": "image_upload_outcome_unknown"}, "decision": {"const": "verified_not_applied"}}},
		  {"required": ["pending_evidence_sha256"], "properties": {"blocker_code": {"const": "image_upload_outcome_unknown"}, "decision": {"const": "verified_uploaded"}}},
		  {"required": ["observed_hash"], "properties": {"blocker_code": {"const": "downloader_partial_add_requires_reconciliation"}, "decision": {"const": "verified_remote_state"}}},
		  {"required": ["observed_hash"], "properties": {"blocker_code": {"const": "downloader_add_outcome_unknown"}, "decision": {"const": "verified_remote_state"}}}
		]
	  },
...
      "RetorrentSummaryArtifact": {
        "type": "object", "additionalProperties": false,
        "required": ["artifact_id", "kind", "storage_path", "size_bytes", "sha256"],
        "properties": {
          "artifact_id": {"type": "string", "format": "uuid"}, "kind": {"type": "string"},
          "storage_path": {"type": "string"}, "mime_type": {"type": "string"},
          "size_bytes": {"type": "integer", "minimum": 0}, "sha256": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"}
        }
      },
      "RetorrentSummary": {
        "type": "object", "additionalProperties": false,
        "required": ["schema_version", "kind", "ok", "status", "job_id", "generated_at", "source", "content", "materials", "target", "seeding", "audit", "blockers", "next_actions", "resume", "summary_file"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "kind": {"type": "string", "const": "upload-assistant.retorrent-summary.v1"},
          "ok": {"type": "boolean", "const": true}, "status": {"type": "string", "const": "complete"},
          "job_id": {"type": "string", "format": "uuid"}, "generated_at": {"type": "string", "format": "date-time"},
          "source": {"type": "object", "required": ["site_code", "torrent_id", "torrent", "rules", "downloader"], "additionalProperties": true},
          "content": {"type": "object", "required": ["file_count", "total_size_bytes", "local_root", "remote_root", "manifest"], "additionalProperties": true},
          "materials": {
            "type": "object", "required": ["metadata", "metadata_enrichment", "media_info", "screenshots", "image_uploads"],
            "properties": {
              "metadata_enrichment": {
                "type": "object", "required": ["required", "tmdb", "ptgen"],
                "properties": {
                  "required": {"type": "boolean"},
                  "tmdb": {
                    "type": "object", "required": ["resolved", "identity", "provider", "adapter", "configuration_sha256", "query_sha256", "artifact_id", "storage_path", "sha256"],
                    "properties": {"resolved": {"type": "boolean"}, "identity": {"type": "object"}, "provider": {"type": "string"}, "adapter": {"type": "string"}, "configuration_sha256": {"type": "string"}, "query_sha256": {"type": "string"}, "artifact_id": {"type": "string"}, "storage_path": {"type": "string"}, "sha256": {"type": "string"}},
                    "additionalProperties": false
                  },
                  "ptgen": {
                    "type": "object", "required": ["resolved", "identity", "provider", "adapter", "configuration_sha256", "query_sha256", "description_sha256", "description_size_bytes", "artifact_id", "storage_path", "sha256"],
                    "properties": {"resolved": {"type": "boolean"}, "identity": {"type": "object"}, "provider": {"type": "string"}, "adapter": {"type": "string"}, "configuration_sha256": {"type": "string"}, "query_sha256": {"type": "string"}, "description_sha256": {"type": "string"}, "description_size_bytes": {"type": "integer", "minimum": 0}, "artifact_id": {"type": "string"}, "storage_path": {"type": "string"}, "sha256": {"type": "string"}},
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              }
            },
            "additionalProperties": true
          },
          "target": {"type": "object", "required": ["site_code", "torrent_id", "details_url", "rules", "duplicate_check", "package", "submitted_torrent", "upload", "downloaded_torrent"], "additionalProperties": true},
					"seeding": {"type": "object", "required": ["downloader_name", "downloader_adapter", "downloader_configuration_sha256", "torrent_hash", "expected_remote_content_path", "apply_labels", "category", "tags", "requirements", "checks", "observation_artifact_id", "observation_sha256"], "additionalProperties": true},
          "audit": {"type": "object", "required": ["artifact_count", "artifacts", "required_gates"], "properties": {"artifact_count": {"type": "integer", "minimum": 1}, "artifacts": {"type": "array", "items": {"$ref": "#/components/schemas/RetorrentSummaryArtifact"}}, "required_gates": {"type": "array", "items": {"type": "string"}}}, "additionalProperties": false},
          "blockers": {"type": "array", "maxItems": 0, "items": {"$ref": "#/components/schemas/Blocker"}},
          "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}},
          "resume": {"type": "object", "required": ["resumable", "reason"], "properties": {"resumable": {"type": "boolean", "const": false}, "reason": {"type": "string", "const": "workflow_complete"}}, "additionalProperties": false},
          "summary_file": {"$ref": "#/components/schemas/RetorrentSummaryArtifact"}
        }
      },
      "JobSummaryValue": {
        "oneOf": [
          {"$ref": "#/components/schemas/RetorrentSummary"},
          {"type": "object", "maxProperties": 0}
        ]
      },
      "Job": {
        "type": "object",
        "required": ["id", "kind", "status", "execution_mode", "input", "blockers", "next_actions", "resume_state", "summary", "created_at", "updated_at"],
        "properties": {
          "id": {"type": "string", "format": "uuid"}, "replay_of_job_id": {"type": "string", "format": "uuid"}, "kind": {"type": "string"}, "status": {"$ref": "#/components/schemas/JobStatus"}, "execution_mode": {"type": "string", "enum": ["auto", "step"]}, "current_step": {"type": "string"},
          "input": {"type": "object"}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}, "resume_state": {"type": "object"}, "summary": {"$ref": "#/components/schemas/JobSummaryValue"},
          "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}, "started_at": {"type": "string", "format": "date-time"}, "finished_at": {"type": "string", "format": "date-time"}
        }
      },
      "JobEnvelope": {
        "type": "object",
        "required": ["ok", "status", "job_id", "blockers", "next_actions", "resume_state", "summary", "job"],
        "properties": {"ok": {"type": "boolean"}, "status": {"$ref": "#/components/schemas/JobStatus"}, "job_id": {"type": "string", "format": "uuid"}, "replay_of_job_id": {"type": "string", "format": "uuid"}, "current_step": {"type": "string"}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}, "resume_state": {"type": "object"}, "summary": {"$ref": "#/components/schemas/JobSummaryValue"}, "job": {"$ref": "#/components/schemas/Job"}}
      },
...
      "MetadataResolveRequest": {"type": "object", "additionalProperties": false, "properties": {"imdb_id": {"type": "string", "pattern": "^tt[0-9]{5,12}$"}, "tmdb_id": {"type": "string", "pattern": "^[0-9]{1,16}$"}, "tmdb_type": {"type": "string", "enum": ["movie", "tv"]}, "douban_id": {"type": "string", "pattern": "^[0-9]{1,16}$"}}, "anyOf": [{"required": ["imdb_id"]}, {"required": ["tmdb_id", "tmdb_type"]}, {"required": ["douban_id"]}]},
...
      "Problem": {"type": "object", "required": ["ok", "status", "error", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean", "const": false}, "status": {"type": "string", "const": "failed"}, "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string"}, "detail": {"type": "string"}}}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}}}
    }
  }
}